UL-10 Added account management unit tests conftest.py
Moved all database fixtures here. Other fixtures should
also go here.
